Rename common exception module
