+++ scinoephile/cli/dictionary/dictionary_search_cli.py
@@ -13,7 +13,7 @@ from scinoephile.common.argument_parsing import (
     input_file_arg,
     int_arg,
 )
-from scinoephile.common.exception import ArgumentConflictError
+from scinoephile.common.exceptions import ArgumentConflictError
 from scinoephile.core.cli import ScinoephileCliBase
 from scinoephile.core.dictionaries import DictionaryEntry
 from scinoephile.dictionaries.lookup import (
