+++ scinoephile/cli/eng/eng_process_cli.py
@@ -13,7 +13,7 @@ from scinoephile.common.argument_parsing import (
     int_arg,
     output_file_arg,
 )
-from scinoephile.common.exception import ArgumentConflictError
+from scinoephile.common.exceptions import ArgumentConflictError
 from scinoephile.core.cli import ScinoephileCliBase, read_series, write_series
 from scinoephile.lang.eng.block_review import get_eng_block_reviewed
 from scinoephile.lang.eng.cleaning import get_eng_cleaned
