+++ scinoephile/cli/yue/yue_transcribe_vs_zho_cli.py
@@ -20,7 +20,7 @@ from scinoephile.common.argument_parsing import (
     output_file_arg,
     str_arg,
 )
-from scinoephile.common.exception import ArgumentConflictError, NotAFileError
+from scinoephile.common.exceptions import ArgumentConflictError, NotAFileError
 from scinoephile.common.file import get_temp_file_path
 from scinoephile.core.cli import ScinoephileCliBase, read_series, write_series
 from scinoephile.core.exceptions import ScinoephileError
