+++ scinoephile/cli/yue/yue_translate_vs_zho_cli.py
@@ -16,7 +16,7 @@ from scinoephile.common.argument_parsing import (
     output_file_arg,
     str_arg,
 )
-from scinoephile.common.exception import ArgumentConflictError
+from scinoephile.common.exceptions import ArgumentConflictError
 from scinoephile.core.cli import ScinoephileCliBase, read_series, write_series
 from scinoephile.multilang.yue_zho.translation import (
     YueVsZhoYueHansTranslationPrompt,
