+++ scinoephile/common/__init__.py
@@ -10,7 +10,7 @@ from __future__ import annotations
 from pathlib import Path
 
 from .command_line_interface import CommandLineInterface
-from .exception import (
+from .exceptions import (
     ArgumentConflictError,
     DirectoryExistsError,
     DirectoryNotFoundError,
