+++ scinoephile/common/exceptions.py
