+++ test/common/test_exceptions.py
@@ -1,12 +1,12 @@
 #  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
 #  and distributed under the terms of the BSD license. See the LICENSE file for details.
-"""Tests of common.exception."""
+"""Tests of common.exceptions."""
 
 from __future__ import annotations
 
 import pytest
 
-from scinoephile.common.exception import (
+from scinoephile.common.exceptions import (
     ArgumentConflictError,
     DirectoryExistsError,
     DirectoryNotFoundError,
