+++ test/common/validation/test_val_literal.py
@@ -8,7 +8,7 @@ from typing import Literal
 
 import pytest
 
-from scinoephile.common.exception import ArgumentConflictError
+from scinoephile.common.exceptions import ArgumentConflictError
 from scinoephile.common.validation import val_literal
 
 type Color = Literal["red", "green", "blue"]
